Fix bug in Q2 + remove compiler warning

+++ trabalho-pratico/src/driver.c
@@ -17,7 +17,7 @@ typedef struct driver
     char *license_plate;
     char *city;
     unsigned short account_creation;
-    double average_rating;
+    float average_rating;
     unsigned short total_rides;
     unsigned short latest_ride;
     bool account_status;
@@ -90,22 +90,22 @@ bool get_driver_account_status(Driver driver)
     return driver->account_status;
 }
 
-unsigned short get_driver_total_rides (Driver driver)
+unsigned short get_driver_total_rides(Driver driver)
 {
     return driver->total_rides;
 }
 
-double get_driver_average_rating (Driver driver)
+float get_driver_average_rating(Driver driver)
 {
     return driver->average_rating;
 }
 
-unsigned short get_driver_latest_ride (Driver driver)
+unsigned short get_driver_latest_ride(Driver driver)
 {
     return driver->latest_ride;
 }
 
-void set_driver_latest_ride (Driver driver, unsigned short new_latest_ride)
+void set_driver_latest_ride(Driver driver, unsigned short new_latest_ride)
 {
     if (driver->latest_ride < new_latest_ride)
         driver->latest_ride = new_latest_ride;
@@ -113,7 +113,7 @@ void set_driver_latest_ride (Driver driver, unsigned short new_latest_ride)
 
 void set_driver_average_rating (Driver driver, unsigned short driver_score)
 {
-    double new_driver_average_rating = (driver->average_rating * driver->total_rides + driver_score) / (driver->total_rides + 1);
+    float new_driver_average_rating = (driver->average_rating * driver->total_rides + driver_score) / (driver->total_rides + 1);
 
     driver->average_rating = new_driver_average_rating;
 }
